+++ gateway.tf
@@ -6,10 +6,11 @@ resource "aws_api_gateway_rest_api" "main" {
   minimum_compression_size = var.api_gateway_minimum_compression_size
   tags                     = module.labels.tags
 
-  binary_media_types = [
+  binary_media_types = concat([
     "application/zip",
-    "application/octet-stream"
-  ]
+    "application/octet-stream",
+    "*/*",
+  ], var.api_gateway_customizations_binary_types)
 
   endpoint_configuration {
     types = ["EDGE"]
@@ -79,10 +80,11 @@ resource "aws_api_gateway_method" "root" {
 }
 
 resource "aws_api_gateway_integration" "root" {
-  rest_api_id = aws_api_gateway_rest_api.main.id
-  resource_id = aws_api_gateway_rest_api.main.root_resource_id
-  http_method = aws_api_gateway_method.root.http_method
-  type        = "MOCK"
+  rest_api_id          = aws_api_gateway_rest_api.main.id
+  resource_id          = aws_api_gateway_rest_api.main.root_resource_id
+  http_method          = aws_api_gateway_method.root.http_method
+  timeout_milliseconds = var.api_gateway_timeout_milliseconds
+  type                 = "MOCK"
   request_templates = {
     "application/json" = jsonencode({ statusCode : 404 })
   }
@@ -103,6 +105,415 @@ resource "aws_api_gateway_integration_response" "root" {
   resource_id = aws_api_gateway_rest_api.main.root_resource_id
   http_method = aws_api_gateway_method.root.http_method
   status_code = "404"
+}
+## /isolation
+resource "aws_api_gateway_resource" "isolation" {
+  rest_api_id = aws_api_gateway_rest_api.main.id
+  parent_id   = aws_api_gateway_rest_api.main.root_resource_id
+  path_part   = "isolation"
+}
+resource "aws_api_gateway_method" "isolation_get" {
+  rest_api_id      = aws_api_gateway_rest_api.main.id
+  resource_id      = aws_api_gateway_resource.isolation.id
+  http_method      = "GET"
+  authorization    = "NONE"
+  api_key_required = false
+}
+
+resource "aws_api_gateway_integration" "isolation_get_integration" {
+  rest_api_id             = aws_api_gateway_rest_api.main.id
+  resource_id             = aws_api_gateway_resource.isolation.id
+  http_method             = aws_api_gateway_method.isolation_get.http_method
+  timeout_milliseconds    = var.api_gateway_timeout_milliseconds
+  integration_http_method = "GET"
+  type                    = "AWS"
+  uri                     = format("arn:aws:apigateway:%s:s3:path/%s/isolation/index.html", var.aws_region, aws_s3_bucket.assets.id)
+  credentials             = aws_iam_role.gateway.arn
+}
+resource "aws_api_gateway_method_response" "isolation_get_method_response" {
+  rest_api_id = aws_api_gateway_rest_api.main.id
+  resource_id = aws_api_gateway_resource.isolation.id
+  http_method = aws_api_gateway_method.isolation_get.http_method
+  status_code = "200"
+  response_models = {
+    "application/json" = "Empty"
+  }
+  response_parameters = {
+    "method.response.header.Content-Length"            = false,
+    "method.response.header.Content-Type"              = false,
+    "method.response.header.Cache-Control"             = true,
+    "method.response.header.Pragma"                    = true,
+    "method.response.header.Strict-Transport-Security" = true
+    "method.response.header.X-Frame-Options"           = true
+  }
+}
+
+resource "aws_api_gateway_integration_response" "isolation_get_integration_response" {
+  rest_api_id       = aws_api_gateway_rest_api.main.id
+  resource_id       = aws_api_gateway_resource.isolation.id
+  http_method       = aws_api_gateway_method.isolation_get.http_method
+  selection_pattern = aws_api_gateway_method_response.isolation_get_method_response.status_code
+  status_code       = aws_api_gateway_method_response.isolation_get_method_response.status_code
+  response_parameters = {
+    "method.response.header.Content-Length"            = "integration.response.header.Content-Length",
+    "method.response.header.Content-Type"              = "integration.response.header.Content-Type",
+    "method.response.header.Cache-Control"             = "'no-store'",
+    "method.response.header.Pragma"                    = "'no-cache'",
+    "method.response.header.Strict-Transport-Security" = format("'max-age=%s; includeSubDomains'", var.hsts_max_age)
+  }
+}
+
+
+
+## /isolation/{key+}
+
+resource "aws_api_gateway_resource" "isolation_key" {
+  rest_api_id = aws_api_gateway_rest_api.main.id
+  parent_id   = aws_api_gateway_resource.isolation.id
+  path_part   = "{key+}"
+}
+
+resource "aws_api_gateway_method" "isolation_key_get" {
+  rest_api_id      = aws_api_gateway_rest_api.main.id
+  resource_id      = aws_api_gateway_resource.isolation_key.id
+  http_method      = "GET"
+  authorization    = "NONE"
+  api_key_required = false
+  request_parameters = {
+    "method.request.path.key" = true
+  }
+}
+resource "aws_api_gateway_integration" "isolation_key_get_integration" {
+  rest_api_id             = aws_api_gateway_rest_api.main.id
+  resource_id             = aws_api_gateway_resource.isolation_key.id
+  http_method             = aws_api_gateway_method.isolation_key_get.http_method
+  timeout_milliseconds    = var.api_gateway_timeout_milliseconds
+  integration_http_method = "GET"
+  type                    = "AWS"
+  uri                     = format("arn:aws:apigateway:%s:s3:path/%s/isolation/{key}", var.aws_region, aws_s3_bucket.assets.id)
+  credentials             = aws_iam_role.gateway.arn
+  request_parameters = {
+    "integration.request.path.key" = "method.request.path.key",
+  }
+}
+
+resource "aws_api_gateway_method_response" "isolation_key_get_method_response" {
+  rest_api_id = aws_api_gateway_rest_api.main.id
+  resource_id = aws_api_gateway_resource.isolation_key.id
+  http_method = aws_api_gateway_method.isolation_key_get.http_method
+  status_code = "200"
+  response_models = {
+    "application/json" = "Empty"
+  }
+  response_parameters = {
+    "method.response.header.Content-Length"            = false,
+    "method.response.header.Content-Type"              = false,
+    "method.response.header.Cache-Control"             = true,
+    "method.response.header.Pragma"                    = true,
+    "method.response.header.Strict-Transport-Security" = true
+  }
+}
+
+resource "aws_api_gateway_integration_response" "isolation_key_get_integration_response" {
+  rest_api_id       = aws_api_gateway_rest_api.main.id
+  resource_id       = aws_api_gateway_resource.isolation_key.id
+  http_method       = aws_api_gateway_method.isolation_key_get.http_method
+  selection_pattern = aws_api_gateway_method_response.isolation_key_get_method_response.status_code
+  status_code       = aws_api_gateway_method_response.isolation_key_get_method_response.status_code
+  response_parameters = {
+    "method.response.header.Content-Length"            = "integration.response.header.Content-Length",
+    "method.response.header.Content-Type"              = "integration.response.header.Content-Type",
+    "method.response.header.Cache-Control"             = "'no-store'",
+    "method.response.header.Pragma"                    = "'no-cache'",
+    "method.response.header.Strict-Transport-Security" = format("'max-age=%s; includeSubDomains'", var.hsts_max_age)
+  }
+}
+
+## /admin-ui
+resource "aws_api_gateway_resource" "admin_ui" {
+  rest_api_id = aws_api_gateway_rest_api.main.id
+  parent_id   = aws_api_gateway_rest_api.main.root_resource_id
+  path_part   = "admin-ui"
+}
+resource "aws_api_gateway_method" "admin_ui_get" {
+  rest_api_id      = aws_api_gateway_rest_api.main.id
+  resource_id      = aws_api_gateway_resource.admin_ui.id
+  http_method      = "GET"
+  authorization    = "NONE"
+  api_key_required = false
+}
+
+resource "aws_api_gateway_integration" "admin_ui_get_integration" {
+  rest_api_id             = aws_api_gateway_rest_api.main.id
+  resource_id             = aws_api_gateway_resource.admin_ui.id
+  http_method             = aws_api_gateway_method.admin_ui_get.http_method
+  timeout_milliseconds    = var.api_gateway_timeout_milliseconds
+  integration_http_method = "GET"
+  type                    = "AWS"
+  uri                     = format("arn:aws:apigateway:%s:s3:path/%s/admin-ui/index.html", var.aws_region, aws_s3_bucket.assets.id)
+  credentials             = aws_iam_role.gateway.arn
+}
+
+resource "aws_api_gateway_method_response" "admin_ui_get" {
+  rest_api_id = aws_api_gateway_rest_api.main.id
+  resource_id = aws_api_gateway_resource.admin_ui.id
+  http_method = aws_api_gateway_method.admin_ui_get.http_method
+  status_code = "200"
+
+  response_parameters = {
+    "method.response.header.Content-Length"            = true,
+    "method.response.header.Content-Type"              = true,
+    "method.response.header.Cache-Control"             = true,
+    "method.response.header.Pragma"                    = true,
+    "method.response.header.Strict-Transport-Security" = true
+  }
+}
+
+resource "aws_api_gateway_integration_response" "admin_ui_get_integration" {
+  rest_api_id       = aws_api_gateway_rest_api.main.id
+  resource_id       = aws_api_gateway_resource.admin_ui.id
+  http_method       = aws_api_gateway_method.admin_ui_get.http_method
+  selection_pattern = aws_api_gateway_method_response.admin_ui_get.status_code
+  status_code       = aws_api_gateway_method_response.admin_ui_get.status_code
+  response_parameters = {
+    "method.response.header.Content-Length"            = "integration.response.header.Content-Length",
+    "method.response.header.Content-Type"              = "integration.response.header.Content-Type",
+    "method.response.header.Cache-Control"             = "'no-store'",
+    "method.response.header.Pragma"                    = "'no-cache'",
+    "method.response.header.Strict-Transport-Security" = format("'max-age=%s; includeSubDomains'", var.hsts_max_age)
+  }
+}
+
+## /admin-ui/{key+}
+
+resource "aws_api_gateway_resource" "admin_ui_key" {
+  rest_api_id = aws_api_gateway_rest_api.main.id
+  parent_id   = aws_api_gateway_resource.admin_ui.id
+  path_part   = "{key+}"
+}
+
+resource "aws_api_gateway_method" "admin_ui_key_get" {
+  rest_api_id      = aws_api_gateway_rest_api.main.id
+  resource_id      = aws_api_gateway_resource.admin_ui_key.id
+  http_method      = "GET"
+  authorization    = "NONE"
+  api_key_required = false
+  request_parameters = {
+    "method.request.path.key" = true
+  }
+}
+resource "aws_api_gateway_integration" "admin_ui_key_get_integration" {
+  rest_api_id             = aws_api_gateway_rest_api.main.id
+  resource_id             = aws_api_gateway_resource.admin_ui_key.id
+  http_method             = aws_api_gateway_method.admin_ui_key_get.http_method
+  timeout_milliseconds    = var.api_gateway_timeout_milliseconds
+  integration_http_method = "GET"
+  type                    = "AWS"
+  uri                     = format("arn:aws:apigateway:%s:s3:path/%s/admin-ui/{key}", var.aws_region, aws_s3_bucket.assets.id)
+  credentials             = aws_iam_role.gateway.arn
+  request_parameters = {
+    "integration.request.path.key" = "method.request.path.key",
+  }
+}
+
+resource "aws_api_gateway_method_response" "admin_ui_key_get_method_response" {
+  rest_api_id = aws_api_gateway_rest_api.main.id
+  resource_id = aws_api_gateway_resource.admin_ui_key.id
+  http_method = aws_api_gateway_method.admin_ui_key_get.http_method
+  status_code = "200"
+  response_models = {
+    "application/json" = "Empty"
+  }
+  response_parameters = {
+    "method.response.header.Content-Length"            = false,
+    "method.response.header.Content-Type"              = false,
+    "method.response.header.Cache-Control"             = true,
+    "method.response.header.Pragma"                    = true,
+    "method.response.header.Strict-Transport-Security" = true
+  }
+}
+
+resource "aws_api_gateway_integration_response" "admin_ui_key_get_integration_response" {
+  rest_api_id       = aws_api_gateway_rest_api.main.id
+  resource_id       = aws_api_gateway_resource.admin_ui_key.id
+  http_method       = aws_api_gateway_method.admin_ui_key_get.http_method
+  selection_pattern = aws_api_gateway_method_response.admin_ui_key_get_method_response.status_code
+  status_code       = aws_api_gateway_method_response.admin_ui_key_get_method_response.status_code
+  response_parameters = {
+    "method.response.header.Content-Length"            = "integration.response.header.Content-Length",
+    "method.response.header.Content-Type"              = "integration.response.header.Content-Type",
+    "method.response.header.Cache-Control"             = "'no-store'",
+    "method.response.header.Pragma"                    = "'no-cache'",
+    "method.response.header.Strict-Transport-Security" = format("'max-age=%s; includeSubDomains'", var.hsts_max_age)
+  }
+}
+
+## /admin
+resource "aws_api_gateway_resource" "admin" {
+  rest_api_id = aws_api_gateway_rest_api.main.id
+  parent_id   = aws_api_gateway_rest_api.main.root_resource_id
+  path_part   = "admin"
+}
+
+resource "aws_api_gateway_authorizer" "admin_authorizer" {
+  name          = "${module.labels.id}-cognito-authorizer"
+  rest_api_id   = aws_api_gateway_rest_api.main.id
+  type          = "COGNITO_USER_POOLS"
+  provider_arns = [aws_cognito_user_pool.admin_user_pool.arn]
+}
+
+
+## /admin/{proxy}
+resource "aws_api_gateway_resource" "admin_proxy" {
+  rest_api_id = aws_api_gateway_rest_api.main.id
+  parent_id   = aws_api_gateway_resource.admin.id
+  path_part   = "{proxy+}"
+}
+
+resource "aws_api_gateway_method" "admin_proxy_options" {
+  rest_api_id      = aws_api_gateway_rest_api.main.id
+  resource_id      = aws_api_gateway_resource.admin_proxy.id
+  http_method      = "OPTIONS"
+  authorization    = "NONE"
+  api_key_required = false
+  request_parameters = {
+    "method.request.path.proxy" = true
+  }
+}
+
+resource "aws_api_gateway_integration" "admin_proxy_options_integration" {
+  rest_api_id             = aws_api_gateway_rest_api.main.id
+  resource_id             = aws_api_gateway_resource.admin_proxy.id
+  http_method             = aws_api_gateway_method.admin_proxy_options.http_method
+  integration_http_method = "OPTIONS"
+  type                    = "HTTP_PROXY"
+  uri                     = format("http://%s/{proxy}", aws_lb.admin.dns_name)
+
+  request_parameters = {
+    "integration.request.path.proxy"              = "method.request.path.proxy",
+    "integration.request.header.X-Routing-Secret" = "'${jsondecode(data.aws_secretsmanager_secret_version.api_gateway_header.secret_string)["header-secret"]}'",
+    "integration.request.header.X-Forwarded-For"  = "'nope'"
+  }
+}
+
+resource "aws_api_gateway_method" "admin_proxy_any" {
+  rest_api_id      = aws_api_gateway_rest_api.main.id
+  resource_id      = aws_api_gateway_resource.admin_proxy.id
+  http_method      = "ANY"
+  authorization    = "COGNITO_USER_POOLS"
+  authorizer_id    = aws_api_gateway_authorizer.admin_authorizer.id
+  api_key_required = false
+  request_parameters = {
+    "method.request.path.proxy" = true
+  }
+}
+
+resource "aws_api_gateway_integration" "admin_proxy_any_integration" {
+  rest_api_id             = aws_api_gateway_rest_api.main.id
+  resource_id             = aws_api_gateway_resource.admin_proxy.id
+  http_method             = aws_api_gateway_method.admin_proxy_any.http_method
+  timeout_milliseconds    = var.api_gateway_timeout_milliseconds
+  integration_http_method = "ANY"
+  type                    = "HTTP_PROXY"
+  uri                     = format("http://%s/{proxy}", aws_lb.admin.dns_name)
+  request_parameters = {
+    "integration.request.path.proxy"              = "method.request.path.proxy",
+    "integration.request.header.X-Routing-Secret" = "'${jsondecode(data.aws_secretsmanager_secret_version.api_gateway_header.secret_string)["header-secret"]}'",
+    "integration.request.header.X-Forwarded-For"  = "'nope'"
+  }
+}
+
+resource "aws_api_gateway_method_response" "admin_proxy_any" {
+  rest_api_id = aws_api_gateway_rest_api.main.id
+  resource_id = aws_api_gateway_resource.admin_proxy.id
+  http_method = aws_api_gateway_method.admin_proxy_any.http_method
+  status_code = "200"
+
+  response_parameters = {
+    "method.response.header.access-control-allow-headers" = true,
+    "method.response.header.access-control-allow-methods" = true,
+    "method.response.header.access-control-allow-origin"  = true
+  }
+}
+
+resource "aws_api_gateway_integration_response" "admin_proxy_any_integration" {
+  rest_api_id = aws_api_gateway_rest_api.main.id
+  resource_id = aws_api_gateway_resource.admin_proxy.id
+  http_method = aws_api_gateway_method.admin_proxy_any.http_method
+  status_code = aws_api_gateway_method_response.admin_proxy_any.status_code
+
+  depends_on = [ 
+    aws_api_gateway_integration.admin_proxy_any_integration
+   ]
+}
+
+
+## /enx/{key+}
+resource "aws_api_gateway_resource" "enx_root" {
+  rest_api_id = aws_api_gateway_rest_api.main.id
+  parent_id   = aws_api_gateway_rest_api.main.root_resource_id
+  path_part   = "enx"
+}
+
+resource "aws_api_gateway_resource" "enx_proxy" {
+  rest_api_id = aws_api_gateway_rest_api.main.id
+  parent_id   = aws_api_gateway_resource.enx_root.id
+  path_part   = "{key+}"
+}
+
+resource "aws_api_gateway_method" "enx_proxy_get" {
+  rest_api_id      = aws_api_gateway_rest_api.main.id
+  resource_id      = aws_api_gateway_resource.enx_proxy.id
+  http_method      = "GET"
+  authorization    = "NONE"
+  api_key_required = false
+  request_parameters = {
+    "method.request.path.key" = true
+  }
+}
+
+resource "aws_api_gateway_integration" "enx_proxy_get_integration" {
+  rest_api_id             = aws_api_gateway_rest_api.main.id
+  resource_id             = aws_api_gateway_resource.enx_proxy.id
+  http_method             = aws_api_gateway_method.enx_proxy_get.http_method
+  timeout_milliseconds    = var.api_gateway_timeout_milliseconds
+  integration_http_method = "GET"
+  type                    = "AWS"
+  uri                     = format("arn:aws:apigateway:%s:s3:path/%s/enx/{key}", var.aws_region, aws_s3_bucket.assets.id)
+  credentials             = aws_iam_role.gateway.arn
+  request_parameters = {
+    "integration.request.path.key" = "method.request.path.key",
+  }
+}
+
+resource "aws_api_gateway_method_response" "enx_proxy_get" {
+  rest_api_id = aws_api_gateway_rest_api.main.id
+  resource_id = aws_api_gateway_resource.enx_proxy.id
+  http_method = aws_api_gateway_method.enx_proxy_get.http_method
+  status_code = "200"
+
+  response_parameters = {
+    "method.response.header.Content-Length"            = false,
+    "method.response.header.Content-Type"              = false,
+    "method.response.header.Cache-Control"             = true,
+    "method.response.header.Pragma"                    = true,
+    "method.response.header.Strict-Transport-Security" = true
+  }
+}
+
+resource "aws_api_gateway_integration_response" "enx_proxy_get_integration" {
+  rest_api_id       = aws_api_gateway_rest_api.main.id
+  resource_id       = aws_api_gateway_resource.enx_proxy.id
+  http_method       = aws_api_gateway_method.enx_proxy_get.http_method
+  selection_pattern = aws_api_gateway_method_response.enx_proxy_get.status_code
+  status_code       = aws_api_gateway_method_response.enx_proxy_get.status_code
+  response_parameters = {
+    "method.response.header.Content-Length"            = "integration.response.header.Content-Length",
+    "method.response.header.Content-Type"              = "integration.response.header.Content-Type",
+    "method.response.header.Cache-Control"             = "'no-store'",
+    "method.response.header.Pragma"                    = "'no-cache'",
+    "method.response.header.Strict-Transport-Security" = format("'max-age=%s; includeSubDomains'", var.hsts_max_age)
+  }
 }
 
 ## /api
@@ -125,13 +536,24 @@ resource "aws_api_gateway_method" "api_proxy_options" {
   http_method      = "OPTIONS"
   authorization    = "NONE"
   api_key_required = false
+  request_parameters = {
+    "method.request.path.proxy" = true
+  }
 }
 
 resource "aws_api_gateway_integration" "api_proxy_options_integration" {
-  rest_api_id = aws_api_gateway_rest_api.main.id
-  resource_id = aws_api_gateway_resource.api_proxy.id
-  http_method = aws_api_gateway_method.api_proxy_options.http_method
-  type        = "MOCK"
+  rest_api_id             = aws_api_gateway_rest_api.main.id
+  resource_id             = aws_api_gateway_resource.api_proxy.id
+  http_method             = aws_api_gateway_method.api_proxy_options.http_method
+  integration_http_method = "OPTIONS"
+  type                    = "HTTP_PROXY"
+  uri                     = format("http://%s/{proxy}", aws_lb.api.dns_name)
+
+  request_parameters = {
+    "integration.request.path.proxy"              = "method.request.path.proxy",
+    "integration.request.header.X-Routing-Secret" = "'${jsondecode(data.aws_secretsmanager_secret_version.api_gateway_header.secret_string)["header-secret"]}'",
+    "integration.request.header.X-Forwarded-For"  = "'nope'"
+  }
 }
 
 resource "aws_api_gateway_method" "api_proxy_any" {
@@ -149,9 +571,10 @@ resource "aws_api_gateway_integration" "api_proxy_any_integration" {
   rest_api_id             = aws_api_gateway_rest_api.main.id
   resource_id             = aws_api_gateway_resource.api_proxy.id
   http_method             = aws_api_gateway_method.api_proxy_any.http_method
+  timeout_milliseconds    = var.api_gateway_timeout_milliseconds
   integration_http_method = "ANY"
   type                    = "HTTP_PROXY"
-  uri                     = "http://${aws_lb.api.dns_name}/{proxy}"
+  uri                     = format("http://%s/{proxy}", aws_lb.api.dns_name)
   request_parameters = {
     "integration.request.path.proxy"              = "method.request.path.proxy",
     "integration.request.header.X-Routing-Secret" = "'${jsondecode(data.aws_secretsmanager_secret_version.api_gateway_header.secret_string)["header-secret"]}'",
@@ -164,6 +587,12 @@ resource "aws_api_gateway_method_response" "api_proxy_any" {
   resource_id = aws_api_gateway_resource.api_proxy.id
   http_method = aws_api_gateway_method.api_proxy_any.http_method
   status_code = "200"
+
+  response_parameters = {
+    "method.response.header.access-control-allow-headers" = true,
+    "method.response.header.access-control-allow-methods" = true,
+    "method.response.header.access-control-allow-origin"  = true
+  }
 }
 
 resource "aws_api_gateway_integration_response" "api_proxy_any_integration" {
@@ -192,9 +621,10 @@ resource "aws_api_gateway_integration" "api_settings_get_integration" {
   rest_api_id             = aws_api_gateway_rest_api.main.id
   resource_id             = aws_api_gateway_resource.api_settings.id
   http_method             = aws_api_gateway_method.api_settings_get.http_method
+  timeout_milliseconds    = var.api_gateway_timeout_milliseconds
   integration_http_method = "GET"
   type                    = "AWS"
-  uri                     = "arn:aws:apigateway:${var.aws_region}:s3:path/${aws_s3_bucket.assets.id}/settings.json"
+  uri                     = format("arn:aws:apigateway:%s:s3:path/%s/settings.json", var.aws_region, aws_s3_bucket.assets.id)
   credentials             = aws_iam_role.gateway.arn
 }
 
@@ -250,9 +680,10 @@ resource "aws_api_gateway_integration" "api_settings_exposures_get_integration" 
   rest_api_id             = aws_api_gateway_rest_api.main.id
   resource_id             = aws_api_gateway_resource.api_settings_exposures.id
   http_method             = aws_api_gateway_method.api_settings_exposures_get.http_method
+  timeout_milliseconds    = var.api_gateway_timeout_milliseconds
   integration_http_method = "GET"
   type                    = "AWS"
-  uri                     = "arn:aws:apigateway:${var.aws_region}:s3:path/${aws_s3_bucket.assets.id}/exposures.json"
+  uri                     = format("arn:aws:apigateway:%s:s3:path/%s/exposures.json", var.aws_region, aws_s3_bucket.assets.id)
   credentials             = aws_iam_role.gateway.arn
 }
 
@@ -307,9 +738,10 @@ resource "aws_api_gateway_integration" "api_settings_language_get_integration" {
   rest_api_id             = aws_api_gateway_rest_api.main.id
   resource_id             = aws_api_gateway_resource.api_settings_language.id
   http_method             = aws_api_gateway_method.api_settings_language_get.http_method
+  timeout_milliseconds    = var.api_gateway_timeout_milliseconds
   integration_http_method = "GET"
   type                    = "AWS"
-  uri                     = "arn:aws:apigateway:${var.aws_region}:s3:path/${aws_s3_bucket.assets.id}/language.json"
+  uri                     = format("arn:aws:apigateway:%s:s3:path/%s/language.json", var.aws_region, aws_s3_bucket.assets.id)
   credentials             = aws_iam_role.gateway.arn
 }
 
@@ -365,9 +797,10 @@ resource "aws_api_gateway_integration" "api_stats_get_integration" {
   rest_api_id             = aws_api_gateway_rest_api.main.id
   resource_id             = aws_api_gateway_resource.api_stats.id
   http_method             = aws_api_gateway_method.api_stats_get.http_method
+  timeout_milliseconds    = var.api_gateway_timeout_milliseconds
   integration_http_method = "GET"
   type                    = "AWS"
-  uri                     = "arn:aws:apigateway:${var.aws_region}:s3:path/${aws_s3_bucket.assets.id}/stats.json"
+  uri                     = format("arn:aws:apigateway:%s:s3:path/%s/stats.json", var.aws_region, aws_s3_bucket.assets.id)
   credentials             = aws_iam_role.gateway.arn
 }
 
@@ -440,9 +873,10 @@ resource "aws_api_gateway_integration" "api_data_exposures_item_get_integration"
   rest_api_id             = aws_api_gateway_rest_api.main.id
   resource_id             = aws_api_gateway_resource.api_data_exposures_item.id
   http_method             = aws_api_gateway_method.api_data_exposures_item_get.http_method
+  timeout_milliseconds    = var.api_gateway_timeout_milliseconds
   integration_http_method = "GET"
   type                    = "AWS"
-  uri                     = "arn:aws:apigateway:${var.aws_region}:s3:path/${aws_s3_bucket.assets.id}/exposures/{item}"
+  uri                     = format("arn:aws:apigateway:%s:s3:path/%s/exposures/{item}", var.aws_region, aws_s3_bucket.assets.id)
   credentials             = aws_iam_role.gateway.arn
   request_parameters = {
     "integration.request.path.item" = "method.request.path.item"
@@ -602,9 +1036,13 @@ resource "aws_api_gateway_integration_response" "api_healthcheck_head_integratio
 # #########################################
 # API Gateway Deployment
 # #########################################
+locals {
+  gw_stage_description = format("%s-%s", filemd5("${path.module}/gateway.tf"), var.api_gateway_customizations_md5)
+}
+
 resource "aws_api_gateway_deployment" "live" {
   rest_api_id       = aws_api_gateway_rest_api.main.id
-  stage_description = filemd5("${path.module}/gateway.tf")
+  stage_description = local.gw_stage_description
 
   lifecycle {
     create_before_destroy = true
@@ -612,6 +1050,13 @@ resource "aws_api_gateway_deployment" "live" {
 
   depends_on = [
     aws_api_gateway_integration.root,
+    aws_api_gateway_integration.isolation_get_integration,
+    aws_api_gateway_integration.isolation_key_get_integration,
+    aws_api_gateway_integration.admin_ui_get_integration,
+    aws_api_gateway_integration.admin_ui_key_get_integration,
+    aws_api_gateway_integration.admin_proxy_options_integration,
+    aws_api_gateway_integration.admin_proxy_any_integration,
+    aws_api_gateway_integration.enx_proxy_get_integration,
     aws_api_gateway_integration.api_proxy_options_integration,
     aws_api_gateway_integration.api_proxy_any_integration,
     aws_api_gateway_integration.api_settings_get_integration,
@@ -639,13 +1084,68 @@ resource "aws_api_gateway_stage" "live" {
   stage_name    = "live"
   access_log_settings {
     destination_arn = aws_cloudwatch_log_group.api_gateway.arn
-    format          = "[$context.requestTime] \"$context.httpMethod $context.resourcePath $context.protocol\" $context.status $context.responseLength $context.requestId"
+    format          = "[$context.requestTime] \"$context.httpMethod $context.path $context.protocol\" $context.status [$context.identity.userAgent] $context.responseLength $context.requestId"
   }
 
   lifecycle {
     ignore_changes = [
       cache_cluster_size
     ]
+  }
+}
+
+resource "aws_cloudwatch_log_metric_filter" "enx_logo_all_filter" {
+  log_group_name = "${module.labels.id}-gw-access-logs"
+  name = "${module.labels.id}-enxlogoall-filter"
+  pattern = "[time ,method = \"*enx/logo*\", statusCode ,agent=\"*CFNetwork*\"]"
+  metric_transformation {
+    name = "enxlogoall"
+    namespace = "ApiGateway"
+    value = "1"
+  }
+}
+
+resource "aws_cloudwatch_log_metric_filter" "enx_logo_200_filter" {
+  log_group_name = "${module.labels.id}-gw-access-logs"
+  name = "${module.labels.id}-enxlogo200-filter"
+  pattern = "[time ,method = \"*enx/logo*\", statusCode = 200,agent=\"*CFNetwork*\"]"
+  metric_transformation {
+    name = "enxlogo200"
+    namespace = "ApiGateway"
+    value = "1"
+  }
+}
+
+resource "aws_cloudwatch_log_metric_filter" "enx_logo_304_filter" {
+  log_group_name = "${module.labels.id}-gw-access-logs"
+  name = "${module.labels.id}-enxlogo304-filter"
+  pattern = "[time ,method = \"*enx/logo*\", statusCode = 304,agent=\"*CFNetwork*\"]"
+  metric_transformation {
+    name = "enxlogo304"
+    namespace = "ApiGateway"
+    value = "1"
+  }
+}
+
+resource "aws_cloudwatch_log_metric_filter" "enx_logo_settings_filter" {
+  log_group_name = "${module.labels.id}-gw-access-logs"
+  name = "${module.labels.id}-enxlogosettings-filter"
+  pattern = "[time ,method = \"*enx/logo*\", statusCode = 200,agent=\"*Settings*CFNetwork*\"]"
+  metric_transformation {
+    name = "enxlogosettings"
+    namespace = "ApiGateway"
+    value = "1"
+  }
+}
+
+resource "aws_cloudwatch_log_metric_filter" "enx_logo_enbuddy_filter" {
+  log_group_name = "${module.labels.id}-gw-access-logs"
+  name = "${module.labels.id}-enxlogoenbuddy-filter"
+  pattern = "[time ,method = \"*enx/logo*\", statusCode = 200,agent=\"*HealthENBuddy*CFNetwork*\"]"
+  metric_transformation {
+    name = "enxlogoenbuddy"
+    namespace = "ApiGateway"
+    value = "1"
   }
 }
 
@@ -675,4 +1175,17 @@ resource "aws_api_gateway_authorizer" "main" {
   rest_api_id            = aws_api_gateway_rest_api.main.id
   authorizer_uri         = coalesce(join("", aws_lambda_alias.authorizer_live.*.invoke_arn), aws_lambda_function.authorizer.invoke_arn)
   authorizer_credentials = aws_iam_role.authorizer.arn
+}
+
+
+# #########################################
+# Default Responses
+# #########################################
+resource "aws_api_gateway_gateway_response" "test" {
+  rest_api_id   = aws_api_gateway_rest_api.main.id
+  response_type = "DEFAULT_4XX"
+
+  response_parameters = {
+    "gatewayresponse.header.access-control-allow-origin" = "'*'"
+  }
 }