+++ users.tf
@@ -3,10 +3,6 @@ resource "aws_iam_user" "ci_user" {
   tags = module.labels.tags
 }
 
-resource "aws_iam_access_key" "ci_user" {
-  user = aws_iam_user.ci_user.name
-}
-
 resource "aws_iam_user_policy_attachment" "ci_user_ecr" {
   user       = aws_iam_user.ci_user.name
   policy_arn = "arn:aws:iam::aws:policy/AmazonEC2ContainerRegistryFullAccess"
@@ -18,9 +14,7 @@ data "aws_iam_policy_document" "ci_user" {
       "ecs:UpdateService",
       "ecs:DescribeTaskDefinition",
       "ecs:RegisterTaskDefinition",
-      "iam:PassRole",
       "ecs:DescribeServices",
-      "lambda:*",
       "cloudwatch:PutDashboard",
       "cloudwatch:GetDashboard"
     ]
@@ -29,10 +23,77 @@ data "aws_iam_policy_document" "ci_user" {
       "*",
     ]
   }
+  statement {
+    actions = [
+      "s3:ListBucket",
+      "s3:PutObject"
+    ]
+
+    resources = [
+      aws_s3_bucket.assets.arn,
+      format("%s/*", aws_s3_bucket.assets.arn)
+    ]
+  }
+
+  statement {
+    actions = [
+      "route53:GetChange"
+    ]
+    resources = [
+      "*"
+    ]
+  }
 }
 
-resource "aws_iam_user_policy" "ci_user" {
+data "aws_iam_policy_document" "ci_user_lambda" {
+  statement {
+    actions = [
+      "lambda:UpdateFunctionCode",
+      "lambda:ListAliases",
+      "lambda:ListVersionsByFunction",
+      "lambda:UpdateAlias"
+    ]
+
+    resources = [
+      "*",
+    ]
+  }
+}
+
+
+data "aws_iam_policy_document" "ci_user_pass_role" {
+  statement {
+    actions = [
+      "iam:PassRole"
+    ]
+
+    resources = [
+      aws_iam_role.admin_ecs_task_role.arn,
+      aws_iam_role.admin_ecs_task_execution.arn,
+      aws_iam_role.api_ecs_task_role.arn,
+      aws_iam_role.api_ecs_task_execution.arn,
+      aws_iam_role.push_ecs_task_role.arn,
+      aws_iam_role.push_ecs_task_execution.arn
+    ]
+  }
+}
+
+resource "aws_iam_user_policy" "ci_user_general" {
   name   = "${module.labels.id}-ci-user"
   user   = aws_iam_user.ci_user.name
   policy = data.aws_iam_policy_document.ci_user.json
 }
+
+resource "aws_iam_user_policy" "ci_user_lambda" {
+  name   = "${module.labels.id}-ci-user_lambda"
+  user   = aws_iam_user.ci_user.name
+  policy = data.aws_iam_policy_document.ci_user_lambda.json
+}
+
+resource "aws_iam_user_policy" "ci_user_pass_role" {
+  name   = "${module.labels.id}-ci-user_pass_role"
+  user   = aws_iam_user.ci_user.name
+  policy = data.aws_iam_policy_document.ci_user_pass_role.json
+}
+
+